requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
pandas>=2.0.0
Python>=3.12.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9